OSX fix time size

+++ module/owlib/src/c/error.c
@@ -89,7 +89,7 @@ void print_timestamp_(const char * file, int line, const char * func, const char
 	va_end( va ) ;
 
 	//` Now print it out
-	fprintf(stderr, "%ld DEFAULT: %s %ld.%06ld\n", time(NULL), buf, tv.tv_sec, tv.tv_usec);
+	fprintf(stderr, "%ld DEFAULT: %s %ld.%06ld\n", time(NULL), buf, (long int) tv.tv_sec, (long int) tv.tv_usec);
 	fflush(stderr);
 }
